app/routes.py
import flask
from app import db, mail,flask_app
import flask_mail
import os
import jwt
from flask_login import login_user, login_required, logout_user, current_user
from werkzeug.security import generate_password_hash, check_password_hash

from app.main_bp import routes
